requirements.txt
absl-py==2.3.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.5
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.2
